requirements.txt
numpy
scipy
tensorly
absl-py
tqdm